+++ pkg/compute/compute_deploy_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/fastly/cli/pkg/compute"
 	"github.com/fastly/cli/pkg/mock"
 	"github.com/fastly/cli/pkg/testutil"
+	"github.com/fastly/go-fastly/v3/fastly"
 )
 
 func TestDeploy(t *testing.T) {
@@ -429,4 +430,58 @@ func makeDeployEnvironment(t *testing.T, manifestContent string) (rootdir string
 	}
 
 	return rootdir
+}
+
+func createServiceOK(i *fastly.CreateServiceInput) (*fastly.Service, error) {
+	return &fastly.Service{
+		ID:   "12345",
+		Name: i.Name,
+		Type: i.Type,
+	}, nil
+}
+
+func createServiceError(*fastly.CreateServiceInput) (*fastly.Service, error) {
+	return nil, testutil.Err
+}
+
+func deleteServiceOK(i *fastly.DeleteServiceInput) error {
+	return nil
+}
+
+func createDomainError(i *fastly.CreateDomainInput) (*fastly.Domain, error) {
+	return nil, testutil.Err
+}
+
+func deleteDomainOK(i *fastly.DeleteDomainInput) error {
+	return nil
+}
+
+func createBackendError(i *fastly.CreateBackendInput) (*fastly.Backend, error) {
+	return nil, testutil.Err
+}
+
+func deleteBackendOK(i *fastly.DeleteBackendInput) error {
+	return nil
+}
+
+func getPackageIdentical(i *fastly.GetPackageInput) (*fastly.Package, error) {
+	return &fastly.Package{
+		ServiceID:      i.ServiceID,
+		ServiceVersion: i.ServiceVersion,
+		Metadata: fastly.PackageMetadata{
+			HashSum: "2b742f99854df7e024c287e36fb0fdfc5414942e012be717e52148ea0d6800d66fc659563f6f11105815051e82b14b61edc84b33b49789b790db1ed3446fb483",
+		},
+	}, nil
+}
+
+func activateVersionError(i *fastly.ActivateVersionInput) (*fastly.Version, error) {
+	return nil, testutil.Err
+}
+
+func listDomainsError(i *fastly.ListDomainsInput) ([]*fastly.Domain, error) {
+	return nil, testutil.Err
+}
+
+func listBackendsError(i *fastly.ListBackendsInput) ([]*fastly.Backend, error) {
+	return nil, testutil.Err
 }
